Refactor Organization name to avoid confusion of Organization enum (#98)
* Refactor Organization name to avoid confusion of Organization enum

* Refactor Organization name to avoid confusion of Organization enum

+++ src/main/kotlin/com/rainist/collectcard/cardbills/CardBillServiceImpl.kt
@@ -103,7 +103,7 @@ class CardBillServiceImpl(
                             val cardOrganization =
                                 organizationService.getOrganizationByObjectId(request.companyId.value)
                             DateTimeUtil.kstNowLocalDate().minusMonths(
-                                cardOrganization?.maxMonth
+                                cardOrganization?.maxMonth.toLong()
                                     ?: CardBillServiceImpl.DEFAULT_MAX_MONTH
                             )
                                 .let {
